requirements.txt
Django==1.8
Markdown==2.6.2
argparse==1.2.1
django-bootstrap3==5.3.1
django-filter==0.10.0
django-moderation==0.3.6
djangorestframework==3.1.3
wsgiref==0.1.2
